requirements.txt
Django==2.2.5
django-heroku==0.3.1
django-mysql==3.2.0
django-cors-headers==3.1.1
djangorestframework==3.10.3
python-dotenv==0.10.3
gunicorn==19.9.0
httplib2==0.9.2
idna==2.6
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.4.4
netifaces==0.10.4
psutil==5.4.2
pycairo==1.16.2
pycups==1.9.73
pycurl==7.43.0.1
django-froala-editor==3.0.6

